add focusedAreaCenter to FocusedAreaOCRView
