+++ .flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"camera_avfoundation","path":"/Users/yo/.pub-cache/hosted/pub.dev/camera_avfoundation-0.9.13+7/","native_build":true,"dependencies":[]},{"name":"google_mlkit_commons","path":"/Users/yo/.pub-cache/hosted/pub.dev/google_mlkit_commons-0.6.1/","native_build":true,"dependencies":[]},{"name":"google_mlkit_text_recognition","path":"/Users/yo/.pub-cache/hosted/pub.dev/google_mlkit_text_recognition-0.11.0/","native_build":true,"dependencies":["google_mlkit_commons"]}],"android":[{"name":"camera_android","path":"/Users/yo/.pub-cache/hosted/pub.dev/camera_android-0.10.8+14/","native_build":true,"dependencies":["flutter_plugin_android_lifecycle"]},{"name":"flutter_plugin_android_lifecycle","path":"/Users/yo/.pub-cache/hosted/pub.dev/flutter_plugin_android_lifecycle-2.0.17/","native_build":true,"dependencies":[]},{"name":"google_mlkit_commons","path":"/Users/yo/.pub-cache/hosted/pub.dev/google_mlkit_commons-0.6.1/","native_build":true,"dependencies":[]},{"name":"google_mlkit_text_recognition","path":"/Users/yo/.pub-cache/hosted/pub.dev/google_mlkit_text_recognition-0.11.0/","native_build":true,"dependencies":["google_mlkit_commons"]}],"macos":[],"linux":[],"windows":[],"web":[{"name":"camera_web","path":"/Users/yo/.pub-cache/hosted/pub.dev/camera_web-0.3.2+2/","dependencies":[]}]},"dependencyGraph":[{"name":"camera","dependencies":["camera_android","camera_avfoundation","camera_web","flutter_plugin_android_lifecycle"]},{"name":"camera_android","dependencies":["flutter_plugin_android_lifecycle"]},{"name":"camera_avfoundation","dependencies":[]},{"name":"camera_web","dependencies":[]},{"name":"flutter_plugin_android_lifecycle","dependencies":[]},{"name":"google_mlkit_commons","dependencies":[]},{"name":"google_mlkit_text_recognition","dependencies":["google_mlkit_commons"]}],"date_created":"2023-12-06 17:15:19.420742","version":"3.13.5"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"camera_avfoundation","path":"/Users/yo/.pub-cache/hosted/pub.dev/camera_avfoundation-0.9.13+7/","native_build":true,"dependencies":[]},{"name":"google_mlkit_commons","path":"/Users/yo/.pub-cache/hosted/pub.dev/google_mlkit_commons-0.6.1/","native_build":true,"dependencies":[]},{"name":"google_mlkit_text_recognition","path":"/Users/yo/.pub-cache/hosted/pub.dev/google_mlkit_text_recognition-0.11.0/","native_build":true,"dependencies":["google_mlkit_commons"]}],"android":[{"name":"camera_android","path":"/Users/yo/.pub-cache/hosted/pub.dev/camera_android-0.10.8+14/","native_build":true,"dependencies":["flutter_plugin_android_lifecycle"]},{"name":"flutter_plugin_android_lifecycle","path":"/Users/yo/.pub-cache/hosted/pub.dev/flutter_plugin_android_lifecycle-2.0.17/","native_build":true,"dependencies":[]},{"name":"google_mlkit_commons","path":"/Users/yo/.pub-cache/hosted/pub.dev/google_mlkit_commons-0.6.1/","native_build":true,"dependencies":[]},{"name":"google_mlkit_text_recognition","path":"/Users/yo/.pub-cache/hosted/pub.dev/google_mlkit_text_recognition-0.11.0/","native_build":true,"dependencies":["google_mlkit_commons"]}],"macos":[],"linux":[],"windows":[],"web":[{"name":"camera_web","path":"/Users/yo/.pub-cache/hosted/pub.dev/camera_web-0.3.2+2/","dependencies":[]}]},"dependencyGraph":[{"name":"camera","dependencies":["camera_android","camera_avfoundation","camera_web","flutter_plugin_android_lifecycle"]},{"name":"camera_android","dependencies":["flutter_plugin_android_lifecycle"]},{"name":"camera_avfoundation","dependencies":[]},{"name":"camera_web","dependencies":[]},{"name":"flutter_plugin_android_lifecycle","dependencies":[]},{"name":"google_mlkit_commons","dependencies":[]},{"name":"google_mlkit_text_recognition","dependencies":["google_mlkit_commons"]}],"date_created":"2023-12-07 19:32:37.727139","version":"3.13.5"}
