+++ lib/focused_area_ocr_painter.dart
@@ -3,18 +3,18 @@ import 'dart:ui' as ui;
 
 import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
+import 'package:focused_area_ocr_flutter/coordinates_translator.dart';
 import 'package:google_mlkit_text_recognition/google_mlkit_text_recognition.dart';
 
-import 'coordinates_translator.dart';
-
-class TextRecognizerPainter extends CustomPainter {
-  TextRecognizerPainter({
+class FocusedAreaOCRPainter extends CustomPainter {
+  FocusedAreaOCRPainter({
     required this.recognizedText,
     required this.imageSize,
     required this.rotation,
     required this.cameraLensDirection,
     required this.focusedAreaWidth,
     required this.focusedAreaHeight,
+    required this.focusedAreaCenter,
     required this.focusedAreaRadius,
     this.focusedAreaPaint,
     this.unfocusedAreaPaint,
@@ -29,6 +29,7 @@ class TextRecognizerPainter extends CustomPainter {
   final CameraLensDirection cameraLensDirection;
   final double focusedAreaWidth;
   final double focusedAreaHeight;
+  final Offset focusedAreaCenter;
   final Radius focusedAreaRadius;
   final Paint? focusedAreaPaint;
   final Paint? unfocusedAreaPaint;
@@ -65,8 +66,9 @@ class TextRecognizerPainter extends CustomPainter {
 
   /// draw unfocused area
   void _drawUnfocusedArea(Canvas canvas, Size size, RRect focusedRRect) {
+    final Offset deviceCenter = Offset(size.width / 2, size.height / 2);
     final Rect deviceRect = Rect.fromCenter(
-      center: Offset(size.width / 2, size.height / 2),
+      center: deviceCenter,
       width: size.width,
       height: size.height,
     );
@@ -136,10 +138,12 @@ class TextRecognizerPainter extends CustomPainter {
   @override
   void paint(Canvas canvas, Size size) {
     final RRect focusedRRect = RRect.fromLTRBR(
-      (size.width - focusedAreaWidth) / 2,
-      (size.height - focusedAreaHeight) / 2,
-      (size.width - focusedAreaWidth) / 2 + focusedAreaWidth,
-      (size.height - focusedAreaHeight) / 2 + focusedAreaHeight,
+      ((size.width - focusedAreaWidth) / 2) + focusedAreaCenter.dx,
+      ((size.height - focusedAreaHeight) / 2) + focusedAreaCenter.dy,
+      ((size.width - focusedAreaWidth) / 2 + focusedAreaWidth) +
+          focusedAreaCenter.dx,
+      ((size.height - focusedAreaHeight) / 2 + focusedAreaHeight) +
+          focusedAreaCenter.dy,
       focusedAreaRadius,
     );
 
@@ -189,7 +193,7 @@ class TextRecognizerPainter extends CustomPainter {
   }
 
   @override
-  bool shouldRepaint(TextRecognizerPainter oldDelegate) {
+  bool shouldRepaint(FocusedAreaOCRPainter oldDelegate) {
     return oldDelegate.recognizedText != recognizedText;
   }
 }
